clusterizer: When we run out of triangles, pick the next closest vertex
Instead of always following the input triangle stream, we first try to
see if any vertex has a wedge that still has live triangles, and pick
the closest vertex to the center out of those.

It's not clear whether we actually should pick the closest one so this
probably needs to be refined later, but this reduces the number of bad
meshlets we generate for meshes with discontinuities.